requirements/base.txt
asgiref==3.9.1
channels==4.2.2
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
django-model-utils==5.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
Markdown==3.8.2
pillow==11.3.0
PyJWT==2.9.0
python-dotenv==1.1.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
tzdata==2025.2
